moves frechet distance to own module
